+++ lib/flutter_agenda.dart
@@ -2,6 +2,7 @@ library flutter_agenda;
 
 export 'src/models/header.dart';
 export 'src/models/header_logo.dart';
+export 'src/models/headers_position.dart';
 export 'src/models/time_slot.dart';
 export 'src/models/resource.dart';
 export 'src/models/agenda_event.dart';