+++ lib/src/models/agenda_event.dart
@@ -26,17 +26,17 @@ class AgendaEvent {
 
   AgendaEvent({
     required this.title,
-    this.subtitle: "",
+    this.subtitle = "",
     required this.start,
     required this.end,
-    this.padding: const EdgeInsets.all(10),
+    this.padding = const EdgeInsets.all(10),
     this.margin,
     this.onTap,
     this.decoration,
-    this.backgroundColor: const Color(0xFF323D6C),
-    this.textStyle: const TextStyle(
+    this.backgroundColor = const Color(0xFF323D6C),
+    this.textStyle = const TextStyle(
         color: Color(0xFF535353), fontSize: 11, fontWeight: FontWeight.w400),
-    this.subtitleStyle:
+    this.subtitleStyle =
         const TextStyle(fontWeight: FontWeight.w800, color: Color(0xFF363636)),
   }) : assert(end.isAfter(start));
 }