+++ lib/src/models/headers_position.dart
@@ -1,0 +1,4 @@
+enum HeadersPosition {
+  top,
+  bottom,
+}