+++ lib/src/styles/agenda_style.dart
@@ -1,42 +1,44 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_agenda/src/models/header_logo.dart';
 import 'package:flutter_agenda/src/models/time_slot.dart';
+import 'package:flutter_agenda/src/models/headers_position.dart';
 
 class AgendaStyle {
   /// Customize the agenda to match you own UI approach.
   /// by defaut the styles are great but you can change them.
   const AgendaStyle({
-    this.startHour: 0,
-    this.endHour: 24,
-    this.direction: TextDirection.ltr,
-    this.cornerBottom: true,
-    this.cornerRight: true,
-    this.pillarColor: Colors.white,
-    this.cornerColor: Colors.white,
-    this.timelineColor: Colors.white,
-    this.timelineItemColor: Colors.white,
-    this.headSeperator: false,
-    this.pillarSeperator: false,
-    this.mainBackgroundColor: Colors.white,
-    this.decorationLineBorderColor: const Color(0xFFCECECE),
+    this.startHour = 0,
+    this.endHour = 24,
+    this.direction = TextDirection.ltr,
+    this.cornerBottom = true,
+    this.cornerRight = true,
+    this.pillarColor = Colors.white,
+    this.cornerColor = Colors.white,
+    this.timelineColor = Colors.white,
+    this.timelineItemColor = Colors.white,
+    this.headSeperator = false,
+    this.pillarSeperator = false,
+    this.mainBackgroundColor = Colors.white,
+    this.decorationLineBorderColor = const Color(0xFFCECECE),
     this.headBottomBorder = true,
     this.fittedWidth = true,
-    this.timelineBorderColor: const Color(0xFF7F7F7F),
-    this.timeItemTextColor: const Color(0xFF7B7B7B),
-    this.eventRadius: 5,
-    this.timeItemTextStyle: const TextStyle(
+    this.timelineBorderColor = const Color(0xFF7F7F7F),
+    this.timeItemTextColor = const Color(0xFF7B7B7B),
+    this.eventRadius = 5,
+    this.timeItemTextStyle = const TextStyle(
         color: Color(0xFFF999999), fontSize: 11, fontWeight: FontWeight.w300),
-    this.pillarWidth: 200,
-    this.headerHeight: 50,
-    this.timeSlot: TimeSlot.half,
+    this.pillarWidth = 200,
+    this.headerHeight = 50,
+    this.timeSlot = TimeSlot.half,
     this.headerLogo = HeaderLogo.circle,
-    this.timeItemWidth: 70,
-    this.decorationLineHeight: 20,
-    this.decorationLineDashWidth: 4,
-    this.decorationLineDashSpaceWidth: 4,
-    this.eventBorderWidth: 4,
-    this.visibleTimeBorder: true,
-    this.visibleDecorationBorder: true,
+    this.timeItemWidth = 70,
+    this.decorationLineHeight = 20,
+    this.decorationLineDashWidth = 4,
+    this.decorationLineDashSpaceWidth = 4,
+    this.eventBorderWidth = 4,
+    this.visibleTimeBorder = true,
+    this.visibleDecorationBorder = true,
+    this.headersPosition = HeadersPosition.top,
   });
 
   final TextDirection direction;
@@ -138,4 +140,7 @@ class AgendaStyle {
 
   /// visible decoration border
   final bool visibleDecorationBorder;
+  
+  /// position of the headers (top or bottom of the screen)
+  final HeadersPosition headersPosition;
 }